Split device.py into multiple modules
Split firmware version related code into separate submodule

Move channel related code from device to dev.channels

Move ExG related code from device to dev.exg

Move device to dev.base

Move methods and comments around
