+++ pyshimmer/__init__.py
@@ -15,9 +15,11 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 from .bluetooth.bt_api import ShimmerBluetooth
 from .bluetooth.bt_commands import DataPacket
-from .uart.dock_api import ShimmerDock
-from .reader.shimmer_reader import ShimmerReader
+from .dev.base import DEFAULT_BAUDRATE
+from .dev.channels import ChannelDataType, EChannelType
+from .dev.exg import ExGMux, ExGRLDLead, ERLDRef, ExGRegister
+from .dev.fw_version import EFirmwareType
 from .reader.binary_reader import ShimmerBinaryReader
+from .reader.shimmer_reader import ShimmerReader
+from .uart.dock_api import ShimmerDock
 from .util import fmt_hex
-from .device import EChannelType, DEFAULT_BAUDRATE, ChannelDataType, ExGRegister, EFirmwareType, ERLDRef, ExGRLDLead, \
-    ExGMux
